[MIG] l10n_fr_pos_caisse_ap_ip from v16 to v17
