+++ l10n_fr_pos_caisse_ap_ip/readme/CONTRIBUTORS.md
@@ -1,0 +1,2 @@
+- Alexis de Lattre \<<alexis.delattre@akretion.com>\>
+- Pierrick Brun \<<pierrick.brun@akretion.com>\>
