--- l10n_fr_pos_caisse_ap_ip/readme/CONTRIBUTORS.rst
@@ -1,2 +1,0 @@
-* Alexis de Lattre <alexis.delattre@akretion.com>
-* Pierrick Brun <pierrick.brun@akretion.com>
