+++ l10n_fr_pos_caisse_ap_ip/readme/CREDITS.md
@@ -1,0 +1,2 @@
+The development of this module has been financially supported by
+[Camptocamp](https://www.camptocamp.com/).
