--- l10n_fr_pos_caisse_ap_ip/readme/CREDITS.rst
@@ -1,1 +1,0 @@
-The development of this module has been financially supported by `Camptocamp <https://www.camptocamp.com/>`_.
